To plot error based on sample size
This code has parts to plot the error based on sample size for all ML modes used along with PCA. The error values are in the Excel file 'sim_track_err_7Knodes_PCA.xlsx'.

Additionally, it contains the code to plot reconstruction error based on no. of PCs considered and another plot showing the number of PCs used for different sample sizes and the consequent reconstruction error.  The error values for these were gathered individually and stored as an array in the code itself.